% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_cep.R
\name{compare_cep}
\alias{compare_cep}
\title{Compare Two ZIP Codes for Equality}
\usage{
compare_cep(df, col_cep1, col_cep2, strict_check = FALSE)
}
\arguments{
\item{df}{A data frame containing the columns with CEP information.}

\item{col_cep1, col_cep2}{Column name for the first and second CEP.}

\item{strict_check}{Logical indicating whether to perform strict checking on 5-digit CEPs.
If \code{TRUE}, evaluates 5-digit codes; if \code{FALSE}, evaluates only 8-digit codes.
Default is \code{FALSE.}}
}
\value{
A vector of the same length whit 1s and 0s. This indicates the comparison result for each row of CEPs in the data frame.
Returns NA for rows not adhering to the criteria (e.g., not 8 or 5 digits).
}
\description{
This function compares two CEPs for equality based on specific criteria.
}
\details{
This function compares two CEPs for equality based on specific criteria.
It checks whether the CEP in two columns of a data frame match certain patterns (8 or 5 digits).
The \code{strict_check} argument allows for strict checking of 5-digit codes.
If the CEP adhere to the specified criteria, the function returns 1 for a match, 0 for a mismatch,
and NA for values not meeting the criteria.
}
\examples{
df <- data.frame(cep1 = c("12345", "12345678", "12345", "12345", "12345"),
                 cep2 = c("12345", "12345678", "54321", "123", "12345"))

compare_cep(df, "cep1", "cep2")                      # Returns c(NA, 1, NA, NA, NA)
compare_cep(df, "cep1", "cep2", strict_check = TRUE) # Returns c(1, 1, 0, NA, 1)

}
