% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extr_cep.R
\name{extr_cep}
\alias{extr_cep}
\title{Extract CEP from Address}
\usage{
extr_cep(addresses, subsector_as_zip = FALSE)
}
\arguments{
\item{addresses}{A character vector or string containing the address information.}

\item{subsector_as_zip}{Logical. If \code{TRUE}, can extracts a 5-digit sub-sector pattern as CEP, when available. Default is \code{FALSE.}}
}
\value{
A vector of the same length with 8-digit. If specified also returns 5-digit CEP when is found, otherwise NA.
}
\description{
This function extracts an 8-digit or 5-digit CEP from a string representing an address.
}
\details{
This function attempts to extract an 8-digit CEP from a provided address string.
It removes punctuation, converts the address to uppercase, and attempts to find a pattern
matching either an 8-digit CEP or a 5-digit CEP followed by a hyphen and a 3-digit extension.
If the \code{subsector_as_zip} argument is \code{TRUE}, it will attempt to extract a 5-digit sub-sector pattern as a CEP
when an 8-digit CEP is not found.
}
\examples{
library(stringr)
library(stringi)
library(dplyr)

extr_cep("12345-678 Some Street, City, Country")  # Returns "12345678"
extr_cep("No ZIP code in this address")  # Returns NA
extr_cep(c("12345-678 Some Street", "No ZIP code here"))  # Returns "12345678" and NA
extr_cep("Only subsector 12345", subsector_as_zip = TRUE)  # Returns "12345"

}
