% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RefineGeo-package.R
\docType{package}
\name{RefineGeo-package}
\alias{RefineGeo}
\alias{RefineGeo-package}
\title{RefineGeo: Facilitate Address Standardization and Intelligent Selection of Geocoding Coordinates}
\description{
RefineGeo is an R package designed to streamline address standardization and intelligent selection of geocoding coordinates. It enhances data accuracy and precision in location-based analyses by providing robust tools for refining addresses and selecting optimal coordinates based on a proprietary methodology.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/PedroTL/RefineGeo}
  \item Report bugs at \url{https://github.com/PedroTL/RefineGeo/issues}
}

}
\author{
\strong{Maintainer}: Lucas João P. T. T. \email{pedrottlucas@gmail.com}

Other contributors:
\itemize{
  \item FEAC Foundation [funder]
  \item Pontifical Catholic University of Campinas [funder]
}

}
\keyword{internal}
