% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_address.R
\docType{data}
\name{sample_address}
\alias{sample_address}
\title{Sample Addresses from Campinas, São Paulo (SP)}
\format{
A data frame with 99 rows and 5 columns:
\describe{
  \item{address}{Contains address information.}
  \item{cep}{Represents the CEP (equivalent to Zip Code) corresponding to the address.}
  \item{municipality}{Specifies the name of the municipality (Campinas).}
  \item{state}{Two-letter abbreviation representing the state (SP - São Paulo).}
  \item{country}{Specifies the name of the country (Brazil).}
  ...
}
}
\source{
<https://github.com/PedroTL/geocoding-mini-curso-r/tree/main/Bancos%20de%20Dados>
}
\usage{
sample_address
}
\description{
This dataset consists of a randomly selected sample of addresses located in Campinas, São Paulo (SP).
It provides information about different random addresses within the city.
}
\keyword{datasets}
