% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_address.R
\name{clean_address}
\alias{clean_address}
\title{Clean Address Function}
\usage{
clean_address(x)
}
\arguments{
\item{x}{A character vector or string containing the address to be cleaned.}
}
\value{
Cleaned address with standardized formatting.
}
\description{
This function cleans an address by removing punctuation, converting to uppercase,
removing accents, and removing extra whitespace.
}
\details{
This function standardizes the formatting of an address by performing the following steps:
\enumerate{
\item Removes punctuation marks from the address.
\item Converts the entire address to uppercase.
\item Removes accents and special characters.
\item Removes extra whitespace to ensure uniform spacing between words.
The cleaned address is returned with standardized formatting.
}
}
\examples{
library(stringr)
library(stringi)
x <- c("Samplê ADReEss  -   Wíth PonctuatìõNs::")
clean_address(x) # Returns "SAMPLE ADDRESS WITH PONCTUATION"

}
